Update VCR usage and shuffle things around
