+++ Readme.md
@@ -36,7 +36,7 @@ file = File.new(File.join(Rails.root, "config", "avalara.yml"))
 if file.exist?
   begin
     AVALARA_CONFIGURATION = YAML.load_file(path)
-    Avalara.configure do |config|
+    Avalara:API.configure do |config|
       config.username = AVALARA_CONFIGURATION["username"] || abort("Avalara configuration file (#{path}) is missing the username value.")
       config.password = AVALARA_CONFIGURATION["password"] || abort("Avalara configuration file (#{path}) is missing the password value.")
       config.version = AVALARA_CONFIGURATION["version"] if AVALARA_CONFIGURATION.has_key?("version")
