+++ spec/models/avalara_spec.rb
@@ -5,103 +5,16 @@ require "spec_helper"
 describe Avalara do
   maintain_contactology_configuration
 
-  let(:configuration) { Avalara.configuration }
+  let(:configuration) { Avalara::API.configuration }
 
-  describe ".configuration" do
-    it "yields a Avalara::Configuration instance" do
-      Avalara.configuration do |yielded|
-        expect(yielded).to be_kind_of Avalara::Configuration
-      end
-    end
-
-    it "yields the same configuration instance across multiple calls" do
-      Avalara.configuration do |config|
-        Avalara.configuration do |config2|
-          expect(config.object_id).to eq config2.object_id
-        end
-      end
-    end
-
-    it "returns the configuration when queried" do
-      Avalara.configuration do |config|
-        expect(Avalara.configuration.object_id).to eq config.object_id
-      end
-    end
-
-    it "may be explicitly overridden" do
-      configuration = Avalara::Configuration.new
-      expect {
-        Avalara.configuration = configuration
-      }.to change { Avalara.configuration }.to(configuration)
-    end
-
-    it "raises an ArgumentError when set to a non-Configuration object" do
-      expect {
-        Avalara.configuration = "bad"
-      }.to raise_error(ArgumentError)
-    end
-  end
-
-  describe ".endpoint" do
-    it "returns the configuration endpoint" do
-      expect(Avalara.endpoint).to eq configuration.endpoint
-    end
-
-    it "overrides the configuration endpoint" do
-      expect {
-        Avalara.endpoint = "https://example.local/"
-      }.to change { configuration.endpoint }.to("https://example.local/")
-    end
-  end
-
-  describe ".username" do
-    it "returns the configuration username" do
-      configuration.username = "username"
-      expect(Avalara.username).to eq configuration.username
-    end
-
-    it "overrides the configuration username" do
-      expect {
-        Avalara.username = "username"
-      }.to change { configuration.username }.to("username")
-    end
-  end
-
-  describe ".password" do
-    it "returns the configuration password" do
-      configuration.password = "password"
-      expect(Avalara.password).to eq configuration.password
-    end
-
-    it "overrides the configuration password" do
-      expect {
-        Avalara.password = "password"
-      }.to change { configuration.password }.to("password")
-    end
-  end
-
-  describe ".version" do
-    it "returns the configuration version" do
-      configuration.version = "version"
-      expect(Avalara.version).to eq configuration.version
-    end
-
-    it "overrides the configuration version" do
-      expect {
-        Avalara.version = "version"
-      }.to change { configuration.version }.to("version")
-    end
-  end
-
-  describe ".get_tax" do
+  describe ".get_tax", vcr: true do
     let(:doc_date) { Date.parse("January 1, 2012") }
     let(:invoice) { FactoryGirl.build_via_new(:invoice, doc_date: doc_date) }
     let(:request) { Avalara.get_tax(invoice) }
     subject { request }
 
-    context "failure" do
+    context "failure", vcr: { :cassette_name => "get_tax/failure" } do
       let(:invoice) { FactoryGirl.build_via_new(:invoice, customer_code: nil) }
-      use_vcr_cassette "get_tax/failure"
 
       it "rasises an error" do
         expect { subject }.to raise_error(Avalara::ApiError)
@@ -131,8 +44,7 @@ describe Avalara do
       end
     end
 
-    context "success" do
-      use_vcr_cassette "get_tax/success"
+    context "success", vcr: { :cassette_name => "get_tax/success" } do
 
       it { is_expected.to be_kind_of Avalara::Response::Invoice }
